# Faster suppress_non_maximum, add comment explaining implementation