Before:

test suppress::test::bench_suppress_non_maximum_decreasing_gradient   ... bench:       3,090 ns/iter (+/- 966)
test suppress::test::bench_suppress_non_maximum_increasing_gradient   ... bench:       4,776 ns/iter (+/- 165)
test suppress::test::bench_suppress_non_maximum_noise_1               ... bench:       7,054 ns/iter (+/- 740)
test suppress::test::bench_suppress_non_maximum_noise_3               ... bench:       5,661 ns/iter (+/- 2,048)
test suppress::test::bench_suppress_non_maximum_noise_7               ... bench:       5,391 ns/iter (+/- 226)

After:

test suppress::test::bench_suppress_non_maximum_decreasing_gradient   ... bench:       2,548 ns/iter (+/- 603)
test suppress::test::bench_suppress_non_maximum_increasing_gradient   ... bench:       3,119 ns/iter (+/- 251)
test suppress::test::bench_suppress_non_maximum_noise_1               ... bench:       6,595 ns/iter (+/- 1,234)
test suppress::test::bench_suppress_non_maximum_noise_3               ... bench:       4,649 ns/iter (+/- 1,636)
test suppress::test::bench_suppress_non_maximum_noise_7               ... bench:       3,731 ns/iter (+/- 1,035)